Disable hostname verification
"Don't do it in production gaaaah", but even with correct certificates for the right host the dashboard STILL doesn't work. It briefly started, with a completely broken GUI, with a single certificate for the correct host. 2 certificates, for the same host, and it somehow tries to compare null to the current hostname and fails.
More info: https://docs.wso2.com/display/ADMIN44x/Enabling+HostName+Verification#EnablingHostNameVerification-Prerequisites